Q4_work/Openai_agents_sdk_practice/Runner_practice/pyproject.toml
[project]
name = "runner-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
]
